refactor : 컴포넌트 이름 수정 및 디렉토리 위치 변경
