+++ src/components/regist/RegistTemplate.jsx
