+++ src/pages/RegistPage.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 
-import UserTemplate from 'components/common/UserTemplate';
+import RegistTemplate from 'components/regist/RegistTemplate';
 import RegistForm from 'components/regist/RegistForm';
 import Responsive from 'components/common/Responsive';
 
 const RegistPage = () => {
   return (
     <Responsive>
-      <UserTemplate>
+      <RegistTemplate>
         <RegistForm />
-      </UserTemplate>
+      </RegistTemplate>
     </Responsive>
   );
 };
